server/requirements-notebooklm.txt
notebooklm-py